ENSITECH-245 refactored popover component into smaller components
