+++ packages/popover/src/scripts/constants/index.ts
@@ -9,3 +9,18 @@ export const DEFAULT_TRANSITION: TransitionOptions = {
  * at which the arrow can be offset from the center.
  */
 export const MIN_ARROW_SHIFT_SIZE = 75;
+
+export const POSITION_OPTIONS = [
+    'top',
+    'top-start',
+    'top-end',
+    'bottom',
+    'bottom-start',
+    'bottom-end',
+    'right',
+    'right-start',
+    'right-end',
+    'left',
+    'left-start',
+    'left-end',
+] as const;
